Fixed location for setup.py
Looks like our daily job will fail if setup.py
is not where it currently resides (higher folder).

But if kept there, pip install wont work since it wont
be part of the package created and uploaded.

Moving it to both places for now for both items to work.

Note: A restructuring is already in the plans to bring
this 'multi-module' (so-called) structure to one straight
strucutre. In that restucturing setup.py would eventually
be in one place anyways.

Issue-ID: OPTFRA-160
Change-Id: Ib9ff7a273d6cf1e0c00bb74e562f268e8a569d48
Signed-off-by: Ikram Ikramullah (ikram@research.att.com) <ikram@research.att.com>