Update patch set 3
Patch Set 3:

Adding Bharath... 

Bharath, this was my concern with pulling MUSIC code into OOF and things potentially falling through the cracks. Since it is "MUSIC API", someone may mistakenly assume that this code is tested in MUSIC. However, this is HAS specific code, so it should be tested. Please advise.

Patch-set: 3