Update patch set 3
Patch Set 3:

Dileep, if reservation/service.py is not supported in R2, I think it should still have unit tests (unit tests are independent of the context the module is used).

If you will not be able to write unit tests for it, I suggest removing it from ONAP branch (e.g. move it to another folder like test).

Patch-set: 3