Update patch set 3
Patch Set 3:

"One more thing" ... all of the Python-facing Music API and related code (including messaging and so on) was to be factored out into its own project (python-music). I had begun this a while back but was then moved off Music before it could be finished. I would strongly recommend this work be resumed at some point. I have the beginnings of it and would be happy to share and/or contribute to its completion in my infinite spare time. Ping me! :)

Patch-set: 3