All compnents on Single docker image
Made changes to docker settings to make a single
image instead of multiple. Also made changes to the
run and build scripts to work on this single image.

The data component requires certs, keys and ca bundles
settings too. For this, made changes to the run script.
Also, do to this, added the feature of picking the default
values from the current directory if arguments are
not provided to the run-docker.scripts.

For example: sudo ./run-dockers.sh will run without providing
any arguments to it IF the directory from where it is run
has these files in it.

1. conductor.conf -- configuration file
2. log.conf -- log settings files
3. aai_cert.cer
4. aai_key.key
5. aai_ca_bunlde.pem (pem) file

The first two from the above have been created and provided
in the current directory (has/). However, since certs will
be different from env to env, the last three needs to be provided
either by copying in the current direcotry or as the arguments 3, 4 and 5.

The script will give a detailed message on what may be missing and
how an attempt was made to get the default files. Checkt it out by
playing with run-docker.sh file.

Finally, the build script also calls the push script now. Since
there can be no testing of this in ONAP artifcats prod BEFORE being
merged to master, we can only test the push part once merged to master.

Issue-ID: OPTFRA-123
Change-Id: I83c54c97953b71aae18166b937dd89195e705f73
Signed-off-by: Ikram Ikramullah <ikram@research.att.com>