Update patch set 2
Patch Set 2:

(1 comment)

So this 'IS' an AAI cert. Can we not keep the older name to be explicit? Extension may change (assuming certs are tested and data will work with .cer extension as well) - can we not keep the same aai_ca_bundle name for this file? The older file may be overriden (if that is the question in mind).

Patch-set: 2