Update patch set 3
Patch Set 3:

> (1 comment)

It is populated with mandatory requirements(values inside 'hpa-feature', 'architecture' and 'hpa-version'). But since in Policy in the future, the 'hpa-feature-attributes' within the same 'hpa-feature' will be separated into multiple data blocks with single 'hpa-feature-attributes'. And in AAI, it is not changed, so it generate this bug. For details, please find OPTFRA-323.

Patch-set: 3
Reviewer: Gerrit User 2977 <2977@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0