Update patch set 1
Patch Set 1:

(1 comment)

The metrics are defined in a global file and will be initialized when the package is imported and won't throw any errors. I can add more unit tests to verify the counters by verifying using the REGISTRY. Is it ok if I submit those in a separate patch as this doesn't affect the coverage much?

Patch-set: 1