Update patch set 2
Patch Set 2:

This would be more along the lines of how HAS can choose to work with and without HTTPs interfaces with MUSIC. Ideally we want to decouple this from our dependency on Music versions, since Music should independently support operating under both HTTPs and HTTP modes. But for the moment, we can add a config parameter that says whether we need to use HTTPS or HTTP and can add a note that this config is music version dependent until MUSIC supports both options.

Patch-set: 2
Label: Code-Review=0